refactor(guest-agent): extract cli framework behavior
